requirements.txt
absl-py==1.3.0
anyio==3.6.1
astunparse==1.6.3
attrs==22.1.0
beautifulsoup4==4.11.1
bottle==0.12.23
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
CProfileV==1.0.7
cycler==0.11.0
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
filelock==3.8.0
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
gdown==4.5.1
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
grpcio-tools==1.47.0
h11==0.13.0
h5py==3.7.0
htmlmin==0.1.12
httptools==0.5.0
huggingface-hub==0.9.1
idna==3.4
ImageHash==4.3.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
missingno==0.5.1
mmh3==3.0.0
multimethod==1.9
networkx==2.8.7
numpy==1.23.3
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.8.0
packaging==21.3
pandas==1.5.0
pandas-profiling==3.4.0
patsy==0.5.3
phik==0.12.2
Pillow==9.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pymilvus==2.1.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sklearn==0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.20.4
statsmodels==0.13.2
tangled-up-in-unicode==0.2.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
tensorflow-metal==0.6.0
termcolor==2.0.1
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.1
torch-scatter==2.0.9
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.22.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
visions==0.7.5
watchfiles==0.17.0
websockets==10.3
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
